pyproject.toml
[project]
name = "meta-color"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colour-science>=0.4.7",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
]
